Big update; Sharingan now uses subtypes
I left in some debugging code so you can see what's going on in the console when you use the item. Feel free to remove the messages.

Update addInformation with whatever you want. Probably get rid of the generic description.

I changed Sharingan to add slow effect with amplifier of 0, so the weakest slow effect, so that Tsukuyomi would feel more powerful in comparison. Tsukuyomi gives slow level 3.

All the durations are currently set to 100 ticks, so 5 seconds.

Cooldowns is currently 10 seconds for Sharingan, 1 minute for Tsukuyomi.

To add new types, simply give it a name and an ID at the top of the file, then add a case for it in the useSharingan method and define its effect. You won't need to do anything else.