Can now place/replace eyes on right click
Default is to replace generic SHARINGAN only on right click, but you can call the 'grantSharingan' method from anywhere, so if you have an event or objective reached, you can set it to override any kind of Sharingan.

Setting a type to override just means that it will replace that type of eye IF it doesn't find an empty slot first. The replaced eye will be given back to the player, just like Obito could take out his eye and give it to Kakashi.

Currently no way to simply remove an eye; you must have a new eye to replace it with.